+++ LabelPrintTests/Properties/AssemblyInfo.cs
@@ -18,5 +18,5 @@ using System.Runtime.InteropServices;
 [assembly: Guid("71e932d3-0722-4f8d-ad80-4276c61854da")]
 
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.5.570.0")]
-[assembly: AssemblyFileVersion("0.5.570.0")]
+[assembly: AssemblyVersion("0.5.580.0")]
+[assembly: AssemblyFileVersion("0.5.580.0")]