+++ ScalesCore/Models/SqlAuthentication.cs
@@ -2,7 +2,6 @@
 // PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 
 using DataProjectsCore;
-using DataShareCore;
 
 namespace ScalesCore.Models
 {